Copied README from private school repository
I copied over the README from the private repository created through my school for a Web Mining course that I took. I could not fork the repository because it was disabled, but my instructor said that we where allowed to make all of our data and results public. He must have just forgotten to enable the ability to fork it.